requirements.txt
streamlit==1.32.0
requests==2.31.0
pandas==2.2.1
numpy==1.26.4
scikit-learn==1.6.1
nltk==3.8.1
python-dotenv==1.0.1 